chore: split acirgen into sections
