+++ compiler/noirc_evaluator/src/acir/generated_acir.rs
@@ -186,16 +186,12 @@ impl<F: AcirField> GeneratedAcir<F> {
         inputs: &[Vec<FunctionInput<F>>],
         constant_inputs: Vec<F>,
         constant_outputs: Vec<F>,
-        output_count: usize,
-    ) -> Result<Vec<Witness>, InternalError> {
+        outputs: Vec<Witness>,
+    ) -> Result<(), InternalError> {
         let input_count = inputs.iter().fold(0usize, |sum, val| sum + val.len());
         intrinsics_check_inputs(func_name, input_count);
-        intrinsics_check_outputs(func_name, output_count);
+        intrinsics_check_outputs(func_name, outputs.len());
 
-        let outputs = vecmap(0..output_count, |_| self.next_witness_index());
-
-        // clone is needed since outputs is moved when used in blackbox function.
-        let outputs_clone = outputs.clone();
 
         let black_box_func_call = match func_name {
             BlackBoxFunc::AES128Encrypt => BlackBoxFuncCall::AES128Encrypt {
@@ -347,7 +343,7 @@ impl<F: AcirField> GeneratedAcir<F> {
 
         self.push_opcode(AcirOpcode::BlackBoxFuncCall(black_box_func_call));
 
-        Ok(outputs_clone)
+        Ok(())
     }
 
     /// Takes an input expression and returns witnesses that are constrained to be limbs
